Implemented male and female ordinal for catalan and some stuff...
...to eficiently handle it.

This is a new method (rulable()) to be able to read 'indirect' rules
pointed by its name.

This indirect rules can be 'official' parameters (starting with '@') or
user defined variables (which will never collide with any new parameter)
which must start with '$'.

It makes easy to specify different language/variation redirections in
the same rule depending on from which variation it was called.

Improed priv_i2a() method documentation trying to make it easier to
understand (Hope yes).